Move galaxy_context fixture up a level
Move galaxy_context up to tests/conftest.py
so ansible_galaxy_cli tests can use it.
